fix modify image

+++ controllers/Sauce.js
@@ -1,5 +1,6 @@
-//Importation du model sauce.js
+//Importation du model sauce.js , package FS(accès aux fonctions qui nous permettent de modifier le système de fichiers)
 const Sauce = require("../models/Sauce");
+const fs = require('fs');
 
 
 //Logique création de sauce et exportation de celle ci
@@ -36,16 +37,45 @@ exports.getOneSauce = (req, res, next) => {
 
 //Logique modification d'une sauce et exportation de celle ci
 exports.modifySauce = (req, res, next) => {
-  Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet modifié" }))
-    .catch((error) => res.status(400).json({ error }));
+  const sauceObject = req.file ? {
+    ...JSON.parse(req.body.sauce),
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+} : { ...req.body };
+
+delete sauceObject._userId;
+Sauce.findOne({_id: req.params.id})
+    .then((sauce) => {
+        if (sauce.userId != req.auth.userId) {
+            res.status(401).json({ message : 'Not authorized'});
+        } else {
+            Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
+            .then(() => res.status(200).json({message : 'Objet modifié!'}))
+            .catch(error => res.status(401).json({ error }));
+        }
+    })
+    .catch((error) => {
+        res.status(400).json({ error });
+    });
 };
 
 //Logique supression d'une sauce et exportation de celle ci
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet supprimé" }))
-    .catch((error) => res.status(400).json({ error }));
+  Sauce.findOne({ _id: req.params.id})
+       .then(sauce => {
+           if (sauce.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = sauce.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Sauce.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
 };
 
 //Logique récupération des sauces et exportation de celle ci
